+++ src/test/java/racinggame/domain/RacingGameTest.java
@@ -2,7 +2,7 @@ package racinggame.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racinggame.vo.RacingGameInfo;
+import racinggame.dto.RacingGameInfo;
 
 import java.util.List;
 
@@ -15,9 +15,10 @@ class RacingGameTest {
     @Test
     void racingGame() {
         //given
-        String numberOfCar = "5";
-        String numberOfAttempt = "3";
-        RacingGame racingGame = new RacingGame(new RacingGameInfo(numberOfCar, numberOfAttempt));
+        int numberOfCar = 5;
+        int numberOfAttempt = 3;
+        RacingGameInfo racingGameInfo = new RacingGameInfo(String.valueOf(numberOfCar), String.valueOf(numberOfAttempt));
+        RacingGame racingGame = new RacingGame(racingGameInfo);
         Engine engine = () -> true;
 
         //when
