+++ src/test/java/racinggame/dto/RacingGameInfoTest.java
@@ -1,4 +1,4 @@
-package racinggame.vo;
+package racinggame.dto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
